renamed insrument folder to instrument and added Picoscope connection/disconnection from GUI
