+++ Instrument/instrument_manager.py
