+++ Instrument/picoscope_manager_test.py
